Remove un-necessary assignment operator template
The assignment operator accepting a modifiable reference to another
'bdema_ManagedPtr' type is not needed, with the implicit conversion
to 'bdema_ManagedPtr_Ref' operator being the preferred conversion
route, reducing the 'conversion to compatible type' logic to the
minimum number of places.
Improved documentation of the imp. detail class
'bdema_ManagedPtr_Members.