Adopt the 'cookie' model for deleters
Redraft documentation for deleters in terms of using a 'cookie'
rather than a 'factory'.  Update all signatures and parameters
to functions to match.  This turned up a nasty corner case in
the most complex form of template, where the cookie used by the
deleter function may be one of multiple bases of the actual
cookie type passed to the bdema_ManagedPtr method.  Fixed by
the wonders of static cast.
Additionally, further tests to verify aliasing of composite
objects, and to properly test diamond-inheritance composites
were added.