Using -bernotok is not actually possible
bde_build is not actually aware of the fact that
shared libraries may depend on other shared libraries.
Therefore it doesnt link in the other libraries when
building, which causes symbols to be defined.

This is only a problem when building the inner package
library (bsls, bslag, bslscm), since when building the
full bsl package, all symbols are indeed defined.

This is going to be a problem also wqhen building the
other groups, but since were not handling that right now
I will just follow with the easy answer right now