debian packaging using producemake for the RSSUITE
reproduce build_prebuild header install

Not bde_build.pl nor producemake actually support the header
installation the way that is actually used in production.
Reproduced the same strategy used by build_prebuild, which is
basically to find and install all headers inside the "comformant"
packages and install them into 00deployed.

stlport is usually a symlink to the source, but here we make a
copy since the source will not be available in the chroot.

Hopefully, this will enable the building of the rssuite the same
way that is built for production today.