Support for SONAME, SHR_SHLIB_DEPS
Two new variables "SONAME" and "SHR_SHLIB_DEPS" can
be used when building shared libraries so that they
can have the actual lib soname set as well as so that
they can link in other shared libraries.

Note that in the SONAME setting doesnt affect AIX.

Additionally the original changes in AIX were reverted
since it was stablished that the problem is in bde_build
itself that tries to build the shared object for each
package, and some packages may not export symbols.

The correct way of solving this may be specific overrides
in the package opts file, but it may also be the case that
the package-level library building dont actually make sense.