Support for uncommon build modes, without touching headers
By avoiding touching the header files, this set of patches reduces the
noise with as many fixes as possible, without transitively supplying
fixes for dependent components through fixing the header file itself.
The majority of fixes involve adding missing headers, related to use
of the 'bsl' namespace when 'BSL_OVERRIDES_STD' is not defined.