Fix test driver errors when exceptions are disabled
Several test drivers fail to compile due to errors in code blocks protected by #ifdefs
unless exceptions are disabled.  This is not a full patch for this mode, as the Sun
compiler that we test on turns try/catch statements into warnings, rather than errors,
so we have considerably more work to resolve these builds with gcc or HP.  Likewise,
this does not attempt to resolve any runtime test driver failures, only the compile
issues.