Test driver patch for Sun/Microsoft string literals
The test driver assumes for a couple of assertions that two string literals
with identical contents will be coallesced to use the same address.  This is
true on some of our compilers, but not on Sun, and not on Microsoft unless in
a release build.  This issue had limitted impact as most string literals come
from test tables, so we are genuinely passing the same address to most assertions.