Work around exception support in usage example
Demonstrating proctors naturally invovles showing their resilience in the face
of code throwing exceptions.  For a quick workaround, such usage examples are
now guarded with a test for BDE_BUILD_TARGET_EXC.  We guard the whole test,
rather than just the local block, to simplify the code that is copy/pasted
back into the component-level documentation.