Initial test in test driver
This patch quickly adds a test to the test driver for non-polymorphic
classes with virtual bases, so we can at least accurately determine
behavior, and add notes to the component documentation for this corner
case.  Committed now to check out and test on Windows, prior to
completing the set of test scenarios.