Test cases for most anticipated corners
Note that this version will not build on IBM due to the inline definition
of default special members requiring private access to their base classes
having private destructors - even though no objects are instantiated and
these methods should not be used.  While this appears to be a bug in the
IBM compiler, it is probably better to work around the issue as it affects
only this one test driver (so far) and would manifesst in user code long
before it became an issue for this component.