Restore macros for platforms not parsing 64bit literals
Although the OpenGrok search came up clean, it turns out that there are
several BDE test drivers (which are not checked in where OpenGrok can
see them) relying on the old macros to detect platforms that cannot
parse 64bit integer literals.  This macro is restored to bsls_platform.h
only, as BDE is the only client so we can ensure there is no need for a
backwards compatibility macro.  In the process, the macro is renamed to
correctly reflect the poor support for *literals* rather than constants.