Migrate bslstl_ex components back into bslstp
The two bslstl_ex... components were originally part of the code adapted
from STLport.  It is not clear why they ended up in the bslstl package
rather than the bslstp package, but as we are looking towards open
sourcing bsl, we want to be sure that all the STLport code in our tree
ends up in a single package that we retain for Bloomberg internal
compatibility, but do not distribute as part of the open source effort.