Correct usage of headers in bsl component headers
Corrects a couple of headers including a higher-level component than necessary, and the knock on into a
couple of containers that were depending on the transitive include effect.  This is mostly accademic
bookkeeping as the affected headers are indirectly included by container implementations, which will in
turn include the higher level headers that are being removed for redundancy.  This is a QoI cleanup
for when we open source.