Small patches to unbreak integration build
This patch-set contains a small number of bug fixes that will allow
dev-integration builds to succeed again, while completing the remaining
work on the three new CString functor components.  Outstanding work is
completing the set of test data for the operator() tests, and producing
a usage example for each component.