Declare operator== as private for bdef_Function
This patch adds private signatures for operator== and operator!= for bdef_Function.
These private operators are necessary to avoid accidentally converting both
function objects to bool, via the implicit convert-to-bool operator, and then
comparing the bool values.  As the test driver defined an operator== for use
validating the various test cases, the test driver is revised to use a
renamed function for the same purpose, to avoid sending the message that these
function objects are Equality Comparable.  Finally, the only accidental use of
this equality comparison operator that I tracked down in BDE was in the
bael_loggermanagerconfiguration component.  This is component is updated in a
basic way to explicitly use the boolean conversions it was implicitly using
before, but the component should be properly reviewed to confirm this was
the intended semantic in the first place - it looks like the component is
trying to provide a guarantee that it cannot deliver.  That additional
review is the subject of the separate ticket, DRQS 32395212.