Initial implementation of hash table
This first commit for the unordered containers branch provides a full implementation
of the HashTable value semantic class and all the machinery needed to build it.  It
does not yet cover the unordered containers themselves, nor their iterators.  Neither
does it have much in the way of documentation nor test drivers.  Each component
confroms to the basic outline of a BDE component though, and the lack of a test
driver implementation for the existing 'bslstl_hash' component is rectified.