Basic implementation of bslalg_hashutil
This initial implementation supplies the code from bdeu_HashUtil::hash2
as the chosen basic implementation for the initial development work.
The plan is to replace it with a more robust, high-performance hash once
we have measurements to back up the proposed replacement.  In the meantime
this code will allow the development of the unodered containers and thier
supporting framework.  Note that we need to provide a suitably levelized
usage example, and fix up the expected values in the primitive 'bslstl_hash'
test driver - in fact that test driver should be completed using generator
functions to properly test the full range of supported (funcdamental)
types for the component.