Initial implementation of unordered containers
This change set includes the initial implementation of the standard interface
for unordered containers built on top of the previously checked-in infrastructure.
This implementation has a number of issues.  The 'set' test driver is modelled
after the associative 'set' test, and currently fails due to asserting tests
that do not hold for an unordered container.  The 'const' and 'modifiable'
iterators for both map containers are not comparable with each other, with
several strategies being explored to resolve this.  Finally, the multiset
container appears to rely on some undeclared method in bslalg::HashTableUtil.
This issue is still to be investigated.