WIP: Provide local tuple-like type to hold hash table parameters
This change set provides a number of formatting fixes, and ongoing corrections for the facility
wide renaming review, although the 'bslstl' package still does not build.  The key addition in
this patch is the addition of 'HashTable_Parameters' as a component-private class to be used to
hold the parameterized data members of the 'HashTable'.  This allows us to exploit the empty
base optimization without making 'HashTable' itself dependant on a parameterized base class.
This avoids various errors with ambiguities at surprising times, and namespace association with
user-supplied types causing name lookup to look in unexpected places.