WIP: Further cleanup towards a building state
This takes care of the correctly constructing non-copyable data members
when a 'HashTable' object is copied, and some other small tweaks.  We
are closing in on a building state, but still have issues with map
'operator[]' and quite a few places that are still passing raw link
pointers to the translator policy, that now expects to be handed the
dereferenced value from each of those links.  Once we have an elegant
solution for this, we should be done.