Many editorial corrections reported by bdechk
While this branch is waiting for review, I took the liberty of also applying
best practice style-guide corrections, so that the components under review
are brought up to spec for the modern BDE - these are come very old components
that have not been touched in quite some time.
Along the way, a couple of corner case bugs were reported and resolved, mostly
with the lack of use of 'explicit' for single-argument constructors, typically
as they became single argument when default arguments were supplied.  The one
nasty issue was an out-of-bounds array access that did not show on the nightly
builds as it occurred only in 'veryVerbose' mode (or more detailed).