Fix the HashTable copy constructor
The copy constructor for HashTable had a number of issues.  The most obvious
fix is that we are now using specific proctor classes to provide exception
safety, rather than throw/catch.  In addition, the copied HashTable will use
an appropriate number of buckets to provide capacity, rather than copying
number of buckets of the original, which could potentially be much higher if
the original object had a higher 'size' at some point prior to the current
state being copied.  This highlighted the worst bug, where an empty
HashTable that at some prior point had been allocated data would use the
universal static bucket, but claim a higher number of buckets that might
be accessed.  All these issues are now resolved, while the logic of the
copy-loop has been optimized to avoid redundant operations.