Add back-compatibility hash functor for bslstp
As we add the new unordered containers, we are supplying an implementation of
'bsl::hash' using a much better hsah function than the identity function
deployed with the STLport containers.  To minimize distruption of existing
clients of those containers using the default hash functor, provide a compatibility
metafunction to produce a functor with the old behavior for any of the fundamental
types the old functor was specialized for, otherwise defer back to the current
implementation of the 'bsl::hash' template for the key type.