Hasty patch for bsl::hash test driver
This patch fixes the failing test cases for the 'bsl::hash' test driver
by simply calling the hash function we expect the functor to call with
the test value, to be sure we are comparing the functor evaluation
against the same expected result.  The test driver still needs a review
to consider if we still have the correct testing strategy now that the
result of hash function is less predictable than with the identity
function that preceded it, and given that we don't want to replicate
the work of testing the 'bslalg' hash function itself, which should
have appropriate coverage in its own test driver.  This is the last
patch necessary before we have a branch that cleanly runs tests and
so can be merged with dev-integration.