Investigate iterator issues on Sun
This does not get to the bottom of the issue, but standard algorithms
have problems with iterators on Sun unless everything it set up just
right so the constituent parts play nicely together.  Apparently, we
have not done this yet.  Checking in so that we can merge to dev-
integration and see the results on other platforms.  Should not
break broad Sun testing as the issue appears confined to a test
driver rather than component headers.