Remove use of noexcept macro
The unordered map component as the ony one providing a forward-compatibility
macro for the C++11 noexcept exception specifications.  While this might
help look like a C++11 contract, there is no benefit unless we are similarly
prepared to use the noexcept operator to provide strong exception safety on
move operations - which we are not even implementing yet.  As this component
was the only one using the macro, it is simpler to withdraw it than generalize
its use.  Some similar portability mechanism will be introduced across the
whole 'bsl' library at some unspecified point in the future, based on
'bsls_compilerfeatures'.