Check point on hash table documentation
The documentation is not so much done as escaping, so we can get feedback.
This change-set provides several other cleanups, consolidating classes and
methods as appropriate, or breaking out new components if warranted.  Some
known naming issues remain, pending a full review of this class for final
method and parameter names.