First pass at HashTable test case 2
Provides sufficient testing infrastructure to look into providing a proper
boot-strap test case for HashTable.  Inspired by the test driver for
'bslstl_multiset' it covers the basics of hashing each of the special types
provided by the template testing facility, by in turn wrapping the desired
hash functor with a call to the test facility 'getIdentifier' function,
that will provide an integer value for an object of any type used in the
test facility.  While we are, implicitly, always testing a hash of 'int',
the functor suppied to the hash table has the correct interface of hasher
for ValueType, and the testing of specific hash functions lies outside this
component, which must merely test that the table works correctly with the
supplied functor or function.  Therefore, with this wrapper, we can test
any combination of hasher and type that are reasonably needed to provide
full test coverage.