More WIP on HashTable tast driver
The test driver is maturing as the appropriate primary accessors and manipulators
are identified, including a small number of test-driver functions rather than
methods of the class under test itself.  The basic tools are now emerging, but
there remains extensive work simply bringing the standard tests up to speed before
starting on the HashTable specific functionality.  Currently the bootstrap test
case 2 is failing where we are expecting some memory allocation guarantees
(inspired by the associative 'mutliset' test driver) that appear to not hold.
Rather than simply drop the tests, we need to understand why this is failing to
be sure it is not some subtle, unsuspected bug (and the test driver doing its
job.