Nest the 'ImplParameters' class within HashTable
Defining the 'HashTable_Parameters' class template outside the
HashTable class was leading to some ugly code and awkwkard
duplication of type aliases.  As the template parameter list
is exactly the same, there is no potential for reducing template
bloat through shared instantiations, so we simplify by making
the 'ImplParameters' a real nested class, rather than an alias
to an instantiation of the previous class template.