Resolve HashTable/Bidirectional pools creating temporaries
One of the nice things about the new allocator_traits feature is
that it correctly propagates allocators to objects constructed
directly into nodes, avoiding creating temporary objects with the
default allocator, by directly forwarding the constructor
arguments into the node element's constructor call.  This patch
makes the necessary changes to the node pool and hash table
components to correctly deduce arguments and forward at all the
necessary stages to avoid creating a premature temporary object.
It also cleans up the implementation of the for 'insert' methods
on hash table, fixing some as-yet unreported bugs that would
spring a memory leak if a user-supplies hasher/comparator
functor threw at the wrong time.  Note that this issue still
remains on the insert-with-hint method, pending finding the
correct proctor to use to manage the new node returned by the
factory past all possible throw-points.