Update test support functions to match contracts
Note that the hash table test driver now fails, as the contract
on the 'insert' function precludes growing past a rehash trigger,
to simplify testing value-semantic operations and holding back
full 'insert' testing until after rehash is tested.  Unfortunately,
the current test plan does not respect this, so a revised test
plan is needed asap.  This branch is being pushed now, even with
a broken test driver, in order to proceed with the parallel merge
of type traits functionality.