Resolve some HashTable test drivers rehash issues
The current version of the 'insertElement' basic manipulator used to
bootstrap testing will refuse to insert, and return 0, if such an
insertion would cause a rehash.  This patch ensures that the test
cases 2 and 3 allocate sufficient buckets up front to avoid the
need to rehash, and so start passing again.  Test case 4 is untouched,
as we want to merge the version recently completed by Raymond.