Resolve memory issues in test case 2
Resolve the odd error reports of wrong allocations in test case 2,
where the logic was inspired by the ordinary associative container
multiset's test driver.  The issue is that I had dismissed some
growth-strategy logic predicting when an allocation should occur,
thinking there was no such issue for the unordered containers -
and that we would need to add logic to separately track allocations
of the bucket array later.  In fact, the odd growth-strategy
calculation was a precise prediction of the effect of node-pooling
in the table's factory.  The methods have been updated with names
to more clearly reflect what is happening, and now the test passes
with flying colors.  We will now revisit the concerns to be sure
the full range of bootstrap concerns for an unordered container
are met, which are more varied than those of the associative
container that has inspired so far.