Mark HashTable as bitwise-moveable
With all the work to get type-traits up and running again, we forgot
to mark the HashTable itself as potentially bitwise-moveable.  Rather
than deferring the answer to the internal 'Impl' struct, we simply
compute the desired trait value, assuming that the factory is bitwise
moveable if the allocator is bitwise-moveable.