setMaxLoadFactor must have constant complexity
The C++11 standard requires that requests so set the maximum load factor
have constant complexity, so we may not rehash when a low load factor is
requested.  Given the choice of ignoring such requests, adjusting the
requested value to the current load factor (minimal value that woulnd
not rehash) and accepting the request and deferring the rehash, we opt
to defer any rehash to the next insert operation.  The allows the
HashTable to have a well-defined state where loadFactor > maxLoadFactor.