Reapply atomics fixes lost in a rebase
Some of the patches for doc/style issues in atomics were dropped
when rebasing on the new 'master' branch created for open source.
This also re-introduced a bug where we were expecting documentation
explaining the specific exception to the style guide rules for the
PowerPC 32-bit atomics .cpp file.  In addition, a small number of
code formatting issues reported by the current 'bdechk' have been
fixed.  The next patch will deal with 'bdespell' and 'bdeflag'
issues.