unordered set iterators are constant iterators
Correct a minor standard compliance issue, that the iterator types for set-like
containers are constant iterators, and so should always return references-to-
const.  This patch takes the simple approach of defining 'iterator' and
'const_iterator' to be aliases for each other, and similarly for the local
iterators used to traverse buckets.