Remove obsolete hash_* dependencies
A small number of bde, bce and bae components hold a vestigial reference
to the pre-standard 'hash_*' containers.  Often this is nothing more than
doc, or #include-ing an old header that is no longer used by the component.
This patch updates doc, and migrates a test driver (not a public header)
to use the new 'unordered' containers instead.  This change has no public
facing impact other than doc, and reducing the exposure to transitive
includes (retained but guarded by a macro-check for transitive-clean builds).