Reduce excessive spew in non-verbose tests
The test driver for unordered_map in particular was spewing a
lot of text to the console, even when not running the test
driver in 'verbose' mode.  Protecting one method of the
breathing test with a check for 'veryVerbose' solves the worst
of the issues, a more deliberate patch addressing the remaining
issues will follow shortly.