Resolve a number of warnings
Patch to fix the majority of avoidablewarnings in the current test
driver 'bslstl_unorderedmap.t.cpp'.  Similar patches may be wanted
for the other unordered container test drivers, but it is also
possible that a number of the affected functions, called only from
the breathing test, might be eliminated or replaced once we have
a full test for  the whole compoent, and the breathing test can
become a simpler piece of code checking the basic configuration of
a container, rather than trying to invoke every method in some
basic way.