finalize unordered_map breathing test
This iteration of the unordered_map breathing test produces no
warnings (from this component), uses the 'bsl_bsltestutil'
component to support testing and has no dependency on <iostream>.
The only dependency on the native std library comes from the
requirement to test that the 'at' function throws the contractual
exceptions specified by the C++11 standard.  This should be the
model for the breathing test of the other 3 unordered components.