Doc, and fix subtle temporary lifetime issues
Update the documentation of unordered_map to better reflect the
presence of 'max_load_factor' in constructors, assignment and
swap.  Current state is not quite stable, as we are about to
change the semantics for setting load factor back to our preferred
but non-standard design, where setting a max load factor below the
current load factor would for a rehash, violating the constant
complexity constraint.

In the middle of this review noticed a subtle issue with the
range-based insert operation, that would create a dangling
reference to a 'key' trying to work out which bucket to store
the new element in.  The solution is to create the node, and
then work out which bucket to use, and whether or not the node
already exists.  This is slight expense, as we are creating the
same temporary object on the heap, in our node, in the case that
the node potentially is not inserted.  Node creation should be
relatively cheap, on the average, due to the pooling strategy we
are employing.  Plus, we can optimize for the case that the
type used to create the element is actually the same type as the
element with a specific overload, as present in the C++11 class
interface.

Still need to extend this patch beyond 'map' and onto the other
three unordered containers, after a fresh review and checking
the test results in the morning.