Remove an un-necessary declaration
There is no optimization to 'insert' when the template type
parameter is a match for 'ValueType' (in HashTable) so drop
that extra overload.  It turns out the only real optimization
opportunity is for 'insertIfMissing', which genuinely does not
need to create the extra node if the matching key is found
in the container.