Updated unordered_map documentation
Overhaul the unordered_map documentation to properly reflect
use of the max_load_factor, including our choice to ignore
the standard requirement to not rehash on a request to set
a max_load_factor lower than the current load_factor.  Some
significant edits trying to correctly reflect how elements
are created by converting from a (deduced) type that might
not be the same as the unordered map's value_type.