Fixed SAFE_2 checks
A number of assertions for SAFE_2 builds were using
badly spelled macros, so not being tested.  Fix the
spelling of the guard macros, and corrects any code
that was previously not tested and would not compile.
Confirm that all the current test drivers pass with
the SAFE_2 checks enabled (other than the already
broken usage-example check in 'bsls_assert.t.cpp'.