HashTableImpUtil::isWellFormed uses correct hasher
As the C++11 standard does not require hash functors to be
DefaultConstructible, we need to pass the (possibly stateful)
hash functor to the 'isWellFormed' function to determine if
the hash table is in a valid arrangement.  The passing convention
is liable to change if a test driver proves that passing by
const-reference does not work unless the 'operator()' overload
of the functor is also const-qualified - which is *not* required
by the C++11 standard.