Address exception-safety in HashTable
There was a surprising test failure in the 'unordered_set' test
driver, that turned out to be due to memory allocation in SAFE_2
builds of the HashTableImpUtil::rehash function, checking that
it has a well-formed anchor as a post-condition.  In principle,
we have no business adding post-condition checks to our functions,
even in SAFE_2 builds, as that is properly the job of test drivers.
However, it did raise an issue that we are not yet testing with a
user-supplied hash function that may throw, and there is a real bug
that needed fixing in this case.  As we have opted to minimize
memory use and not pay for caching the hash-code with each node,
there is a potential for a rehash opertion to fail in the middle
of re-organizing the list, leaving our container scattered into
two lists, and having two anchors with bad bucket indexes.  The
problem highlighted was that we are also leaking the 'spare'
bucket list when this happens too.  As we have no ability to
guarantee we can rebuild an index correctly, we cannot roll back
our operation and provide a strong exception safety guarantee.
The best we can do for the basic guarantee is to clear the
whole HashTable.  To guarantee this is possible, we need to
have either a single list guaranteed to be returned from the
ImpUtil rehash function, even if an exception is thrown, or
to pass the address of the original anchor's root node, rather
than the list root address itself, so that we can update the
original address each time we splice an element from one list
to the other.  We opt to guarantee the whole list transfers
from the old root to the new anchor, guaranteed, even if an
exception is thrown.  However, while the list remains
traversible, there will be no guarantee of ordering and the
bucket index in the anchor may be corrupt if exiting via an
exception.  This provides sufficient guarantee for the hash
table to manage its own cleanup.  As this requires two very
specific proctor classes, I have chosen to implement them
as local classes, although this is far from idiomatic BDE
style, so they may turn into component-private classes
yet, pending a code review.  To complete the test fixes, the
SAFE_2 post-condition check is removed, and added to
appropriate places in the HashTableImpUtil test driver, and
also added as a class invariant check (in SAFE_2 mode) in
the destructor of 'HashTable'.