Work around 'swap' bugs in native libraries
The 'swap' implementation on AIX uses a comma operator, so
is not compatible with our degenerate functor wrapper.
Likewise, the Microsoft implementation of 'swap' prior to
VC2010 takes the address of both arguments to detect and
optimize away self-swapping.  Both libraries are tripped
up by the HashTable test driver, so we provide a specific
'swap' overload for the tricky type that can be found via
ADL.
A future revision of this test driver should be able to
configure whether the ADL-swap is available, as some of
the early boot-strap tests should not require a 'swap'
function, and it is required for only a limitted number
of (unordered container) operations in the C++11 standard.