Eliminate warning: 'objPtr' may be used uninitialized
The warning arises from a poor static analysis not detected
that the catch-all default at the end of a switch statement
cannot be triggered.  As the 'default' case already includes
an ASSERTV to report the error, directly return from the
switch statement rather progressing with a known bad state
through the rest of the test case.