Extend test support for non-swappable functors
Expand test coverage for 'degenerate' functors to include wrappers
that are not Swappable.  Note that some tests will require support
for 'swap', but it is a minority rather than the majority (and
should be called out it the method-under-test's requirements).
To achieve this with the minimum of fuss, simply add an extra
template argument, a 'bool' value, to the DegenerateClass wrapper,
and do not provide a 'swap' free-function overload for the
'false' specialization.  Default the template parameter to 'true'
so that existing test cases are not changed.