HashTable constructor can specify intial maxLoadFactor
The HashTable class has no need to be bound by the current standard
requirement that 'maxLoadFactor' default to 1.0, and indeed it
seems better, from a re-usable component design, to allow the user
to specify this as a constructor argument.  Doing so will also
simplify bootstrap testing, as we can set various inital max
load factors without having to test the 'setMaxLoadFactor' method
which may trigger a rehash.

Note that this initial commit compiles and runs, without breaking
the test drivers of the 4 unordered containers.  However, the
HashTable test driver itself needs to be revisitted to carefully
recompute a number of expected results for test cases that assumed
the 'maxLoadFactor' would always be 1.0, and so computed memory
consumption, bucket counts etc. based on that assumption.