Add missing bsltf functionality
There are two important pieces of missing functionality
in the template testing facility, exposed while trying to
test the bslstl::HashTable class template.  First, there
are a small number of test types missing their 'debugprint'
overload, which means their value cannot be displayed in
an ASSERTV macro - which is a big problem when passing the
'ALL' set of types to a test runner.

Added a proper test-driver support for this feature using
the RUN_EACH_TYPE macro with TEST_TYPES_ALL to ensure that
'debugprint', 'P_' and 'ASSERTV' are compatible with all
test types used by this component.  This required one final
tweak to the test driver to ensure that the 'debugprint'
overloads in this component to not hide those from the
'bsls' namespace, which are important if we want to print
pointers, especially function pointers, and pointer-to-
members.

The other vital part missing from this component is some
way to use the NonEqualityComparable type in ASSERTV
expressions.  The solution adopted here is to add a
test-support macro to be (renamed to something shorter
and) used in place of '==' in test drivers whose test
cases are driven by the template test facility, when
trying to compare objects of the templated testing
type.  This allows us to provide special case support
for the awkward type.