Ongoing HashTable test driver work
This patch generalizes many of the test cases to
handle a variety of 'maxLoadFactor' values, that
are set by the updated constructor for this class.
Support was also broadened in TestCase2, the only
test case that is remotely complete at this point,
to handle awkward element types through the templat
test facility.  One unusual feature is that it
appears we need to specialize the 'bsl::equal_to'
class template directly in this test driver, as
we do not want to couple the NonEqualityComparable
test type with 'bsl::equal_to' in either component.
As this specialization should not actually be used
in any header files, it is a reasonable risk to
place the specialization directly in the test
driver, and assume there will be no One Definition
Rule violations.

We note that the final added test scenario, acting
as a substitute for maps, does not quite work yet
as advertised.  There are two issues:
1. The 'extractKey' function returns an 'int' by
   value, rather than by 'const int&', as it
   must compute the value.  It does not have an
  actual 'int' object to return a reference to.
Unfortunately, other layers of the implementation
components expect to pass on this result, usually
a reference, as a reference.  While a temporary
will bind to the 'const &' arguments in the next
layers, they are passing a reference to stale
memory that may be overwritten, and optimized
builds give some very interesting failure modes
taking advantage of this undefined behavior.

2. The insert contiguous elements test is
   failing, but only in certain configurations.
It is assumed likely that the keys extracted
from some types in the test driver may compare
equal, when using the computed key rather than
the whole object as value.  In such a case,
we need a more subtle test for contiguous
ranges.