Fix for computed key in HashTable tests
To preserve a live reference, we create a cache of the 16
most recently retreived key computations as a static array
within the function.  We know from the nature of the test
driver that there will never be more than 16 live references
to computed results at any one time - it is highly unlikely
there will ever be as many as 5, so 16 is a conservative
size.  This allows us to return a reference from this
array, and so avoid the problems of supporting a returned
temporary value through multiple layers of our APIs.

Next problem to investigate is the broken copy constructor.