Support hash table functor with non-const operator()
In order to support careless users supplying a hash or key-comparison
functor with a function-call operator lacking the 'const' qualifier,
we add an extra layer of indirection with a functor adapter holding
a mutable functor data member.  This eliminates the potential use of
the empty base class optimization, which we hope to restore at some
point in the future by detecting that the function call operator
selected by overload resolution is actually declared as 'const'.
Unfortunately, the template metaprograms that can detect this
all appear to rely on C++11 features, an it may be that we need the
generalized SFINAE rules of C++11 to deliver this.  Certainly it is
proving difficult to write such a metaprogram with even one of our
currently supported C++03 compilers, never mind portable across all
of them.
We are considering providing a specialization for 'bsl::hash' as
the overwhelmingly common case, and then requiring users that
specialize this template for their own types to declare the function
call operator as const, which may be a slight restriction compared
to the standard, but is a much lower barrier than requiring that
all hash functors declare a 'const' operator.