Fix serious errors in HashTable copy constructor
The copy constructor for bslstl::HashTable appears to perform
well in simple cases, but quickly turned up several tricky
errors when varying the maxLoadFactor of the object being
copied (so that the copy used a different number of buckets)
and when testing for basic exception safety, as a number of
workarounds for earlier bugs had subtly introduced bugs
related to guards and protoctors no longer guarding what
they thought they were guarding.  Thankfully, a fairly
detailed test driver exposed these issues, and this version
of the branch should again be ready to merge for integration,
passing all of its current tests.  Note, however, that the
test driver is still not complete.